+++ api/v1alpha1/mesh_types.go
@@ -54,6 +54,10 @@ type MeshSpec struct {
 type MeshStatus struct {
 }
 
+// Markers for generating manifests for Mesh webhooks.
+//+kubebuilder:webhook:path=/mutate-mesh,mutating=true,failurePolicy=fail,sideEffects=None,groups=greymatter.io,resources=meshes,verbs=create;update,versions=v1alpha1,name=mutate-mesh-webhook.greymatter.io,admissionReviewVersions={v1,v1beta1}
+//+kubebuilder:webhook:path=/validate-mesh,mutating=false,failurePolicy=fail,sideEffects=None,groups=greymatter.io,resources=meshes,verbs=create;update;delete,versions=v1alpha1,name=validate-mesh-webhook.greymatter.io,admissionReviewVersions={v1,v1beta1}
+
 // +kubebuilder:object:root=true
 // +kubebuilder:subresource:status
 // +kubebuilder:resource:scope=Namespaced
